--- a/core/src/modules/ops/profile.ts
+++ b/core/src/modules/ops/profile.ts
@@ -8,7 +8,7 @@ import { Event } from "../../models/Event";
 import { GroupMember } from "../../models/GroupMember";
 import { Log } from "../../models/Log";
 import { api } from "actionhero";
-import { Op, OrderItem, WhereAttributeHash } from "sequelize";
+import { Op, OrderItem, WhereAttributeHash, QueryTypes } from "sequelize";
 import { waitForLock } from "../locks";
 import { ProfilePropertyOps } from "./profileProperty";
 import { CLS } from "../../modules/cls";
@@ -686,54 +686,19 @@ export namespace ProfileOps {
    * Task `profile:completeImport` will be enqueued for each Profile.
    */
   export async function makeReady(limit = 100, toExport = true) {
-    // let profiles: Profile[] = await api.sequelize.query(
-    //   `SELECT
-    //       profiles.id as id
-    //     FROM profiles
-    //     JOIN
-    //       "profileProperties" on "profileProperties"."profileId" = profiles.id
-    //     WHERE
-    //       profiles.state = 'pending'
-    //     GROUP BY profiles.id
-    //     HAVING
-    //       MAX("profileProperties".state) = 'ready'
-    //       AND COUNT(DISTINCT "profileProperties".state) = 1
-    //     LIMIT ${limit}
-    //     ;`,
-    //   { type: QueryTypes.SELECT }
-    // );
-
-    let profiles = await Profile.findAll({
-      subQuery: false,
-      attributes: ["id"],
-      where: { state: "pending" },
-      include: [{ model: ProfileProperty, required: true, attributes: [] }],
-      group: "Profile.id",
-      limit,
-      having: {
-        [Op.and]: [
-          api.sequelize.where(
-            api.sequelize.fn(
-              "MAX",
-              api.sequelize.col("profileProperties.state")
-            ),
-            "=",
-            "ready"
-          ),
-          api.sequelize.where(
-            api.sequelize.fn(
-              "COUNT",
-              api.sequelize.fn(
-                "DISTINCT",
-                api.sequelize.col("profileProperties.state")
-              )
-            ),
-            "=",
-            1
-          ),
-        ],
-      },
-    });
+    let profiles: Profile[] = await api.sequelize.query(
+      `
+    SELECT "id" from "profiles" where "state" = 'pending'
+    EXCEPT
+    SELECT DISTINCT("profileId") FROM "profileProperties" WHERE "state" = 'pending'
+    LIMIT ${limit}
+    ;
+    `,
+      {
+        type: QueryTypes.SELECT,
+        model: Profile,
+      }
+    );
 
     const updateResponse = await Profile.update(
       { state: "ready" },