This PR attempts to speed up 3 tasks:
* `profiles:checkReady`
* `system:status`
* `runs:updateCounts`
* `profileProperties:sweep`

**Interesting Learnings**
* With larger datasets (~10M Profile Properties) an `EXCLUDE` query is faster than a sub-query, a CTE, or even a join

--- 

Raw Notes

```sql
-- *** runs:updateCounts ***

-- Before: 10s
-- Resolution: CREATE INDEX "imports_created_profile" ON "imports" ("createdProfile")
-- After: 3s
EXPLAIN ANALYZE 
SELECT count(*) AS "count" FROM "imports" AS "Import" WHERE "Import"."creatorId" = 'run_4473c0af-47cb-4506-80f8-b4ecddb33185' AND "Import"."createdProfile" = true;


-- Before: 5s
-- Resolution: CREATE INDEX "profile_properties_raw_value" ON "profileProperties" ("rawValue")
-- After: 1s
EXPLAIN ANALYZE 
SELECT count(DISTINCT("ProfileMultipleAssociationShim"."id")) AS "count" FROM "profiles" AS "ProfileMultipleAssociationShim" INNER JOIN "profileProperties" AS "ProfileProperties_10" ON "ProfileMultipleAssociationShim"."id" = "ProfileProperties_10"."profileId" AND "ProfileProperties_10"."propertyId" = 'ltv' INNER JOIN "profileProperties" AS "ProfileProperties_9" ON "ProfileMultipleAssociationShim"."id" = "ProfileProperties_9"."profileId" AND "ProfileProperties_9"."propertyId" = 'last_purchase_date' INNER JOIN "profileProperties" AS "ProfileProperties_8" ON "ProfileMultipleAssociationShim"."id" = "ProfileProperties_8"."profileId" AND "ProfileProperties_8"."propertyId" = 'last_purchase_category' WHERE ((CAST("ProfileProperties_10"."rawValue" AS FLOAT) > '100') AND (CAST("ProfileProperties_9"."rawValue" AS BIGINT) > 1617739845527 AND CAST("ProfileProperties_9"."rawValue" AS BIGINT) <= 1622923845527) AND (CAST("ProfileProperties_8"."rawValue" AS TEXT) = 'Automotive'));

--- *** system:status ***

-- Before: 7s
-- Resolution: CREATE INDEX "imports_exported_at" ON "imports" ("exportedAt")
-- After: 0.3s
EXPLAIN ANALYZE
SELECT count(*) AS "count" FROM "imports" AS "Import" WHERE "Import"."exportedAt" IS NULL;

-- Before: 9s
-- Resolution: CREATE INDEX "profile_properties_state" ON "profileProperties" ("state")
-- After: 4s
-- This one is still going to be slow - `SELECT COUNT(DISTINCT("profileId")) from "profileProperties"` takes 4 seconds on its own
EXPLAIN ANALYZE 
SELECT "Property"."sourceId", COUNT(DISTINCT("profileId")) AS "count" FROM "properties" AS "Property" INNER JOIN "profileProperties" AS "ProfileProperties" ON "Property"."id" = "ProfileProperties"."propertyId" AND "ProfileProperties"."state" = 'pending' WHERE "Property"."state" NOT IN ('draft') GROUP BY "sourceId";


--- *** profiles:checkReady ***

-- Before: 30s
-- Resolution: (indexes added above)  
-- After: 17s
-- This one is still going to be slow as we need to compare /all/ the profile properties for pending profiles
EXPLAIN ANALYZE
SELECT "Profile"."id" FROM "profiles" AS "Profile" INNER JOIN "profileProperties" AS "profileProperties" ON "Profile"."id" = "profileProperties"."profileId" WHERE "Profile"."state" = 'pending' GROUP BY "Profile"."id" HAVING (MAX("profileProperties"."state") = 'ready' AND COUNT(DISTINCT("profileProperties"."state")) = 1) LIMIT 1000;

SELECT "profileProperties"."profileId" FROM "profileProperties" INNER JOIN "profiles" AS "Profile" ON "Profile"."id" = "profileProperties"."profileId" WHERE "Profile"."state" = 'pending' GROUP BY "profileProperties"."profileId" HAVING (MAX("profileProperties"."state") = 'ready' AND COUNT(DISTINCT("profileProperties"."state")) = 1) LIMIT 1000;

-- 5 seconds
SELECT id from profiles where state = 'pending' 
EXCEPT 
SELECT distinct("profileId") FROM "profileProperties" where state = 'pending'
LIMIT 1000
;

-- +1 minute
SELECT id from profiles where state = 'pending' AND id NOT IN (SELECT "profileId" FROM "profileProperties" where state = 'pending') LIMIT 1000
;

-- +1 minute
WITH "readyProfileProperties" as (SELECT "profileId" FROM "profileProperties" where state = 'pending') 
SELECT id from profiles where state = 'pending' AND id NOT IN (select "profileId" from "readyProfileProperties") LIMIT 1000
;
```